Split mqtt: forwarder is an isolated class.
